schema splitted
