+++ src/type-defs/index.js
@@ -1,52 +1,7 @@
-const gql = require('graphql-tag');
+const root = require('./root');
+const user = require('./user');
+const hero = require('./hero');
 
-module.exports = gql`
-    enum ROLE {
-        TANK
-        MAGE
-        SUPPORT
-        FIGHTER
-        ASSASSIN
-    }
+const schemaArray = [root, user, hero];
 
-    type User {
-        id: ID!
-        name: String!
-        email: String!
-        password: String!
-        heroes: [Hero]
-    }
-
-    type Hero{
-        id: ID!
-        name: String!
-        role: ROLE!
-    }
-
-    type Query {
-        allUsers: [User]
-        user: User
-    }
-
-    input CreateUserInput {
-        name: String!
-        email: String!
-        password: String!
-    }
-
-    input CreateHeroInput {
-        name: String!
-        role: ROLE!
-    }
-
-    input addHeroInput {
-        name: String!
-    }
-
-    type Mutation {
-        createUser(input: CreateUserInput): User!
-        createHero(input: CreateHeroInput): Hero!
-        addHero(input: addHeroInput): User!
-    }
-`;
-
+module.exports = schemaArray;
